requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.1.0
xgboost==2.1.4
scikit-learn==1.6.1
numpy==2.2.3
scipy==1.15.2
joblib==1.4.2
understat==0.1.12
aiohttp==3.11.14
